+++ src/netcore/models/net-core.model.ts
@@ -1,4 +1,4 @@
-import { AssemblyInfo } from './assembly-info.model';
+import { AssemblyInfo } from './';
 
 export class NetCore extends AssemblyInfo {
     generatePackageOnBuild: boolean = false;