+++ src/netcore/services/utils.service.ts
@@ -1,8 +1,8 @@
 import moment = require('moment');
 import { LoggingLevel } from '../enums';
-import * as models from '../models';
+import * as sharedModels from '../models';
 
-export function transformDates(value: string, regExModel: models.RegEx): string {
+export function transformDates(value: string, regExModel: sharedModels.RegEx): string {
     return value.replace(regExModel.date, (match: string, g1: any, g2: any): string => {
         return moment().format(g1);
     });